Move arithmetic_index into separate file
