+++ libvast/CMakeLists.txt
@@ -337,6 +337,7 @@ set(tests
     test/hash_index.cpp
     test/http.cpp
     test/ids.cpp
+    test/index/arithmetic_index.cpp
     test/iterator.cpp
     test/json.cpp
     test/meta_index.cpp
